feat: ios support

+++ install-pyzule.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
-OS=$(uname)
 ARCH=$(uname -m)
+
+if [[ $ARCH == *"iPhone"* ]]; then
+    OS="iPhone"
+    PATHPREFIX="/var/jb"
+    PYZULEURL="https://raw.githubusercontent.com/TbhLovers/pyzule-ios/main/pyzule-ios.py"
+else
+    OS=$(uname)
+    PYZULEURL="https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/pyzule.py"
+fi
+
 PZ_DIR=${HOME}/.config/pyzule
 
 # check which python we should use
@@ -17,6 +26,8 @@ else
     echo "[!] couldn't find \"python\" nor \"python3\" installed."
     if [ "$OS" == "Linux" ]; then
         echo "[*] try \"sudo apt install python3 python3-pip python3-venv\" or \"sudo pacman -S python python-pip\" depending on your distro."
+    elif [ "$OS" == "iPhone" ]; then
+        echo "[*] try looking for python in your package manager!"
     else
         echo "[*] for installation instructions, head over to python.org !"
     fi
@@ -27,8 +38,11 @@ mkdir -p ${PZ_DIR}
 if [ ! -d ${PZ_DIR}/venv ]; then
     echo "[*] installing required pip libraries.."
     $PYTHON -m venv ${PZ_DIR}/venv > /dev/null
-    ${PZ_DIR}/venv/bin/pip install -U Pillow lief orjson requests &> /dev/null
-elif [ ! -f ${PZ_DIR}/requests_upd ]; then
+
+    if [ ! "$OS" == "iPhone" ]; then
+        ${PZ_DIR}/venv/bin/pip install -U Pillow orjson requests lief &> /dev/null
+    fi
+elif [ ! "$OS" == "iPhone" ] && [ ! -f ${PZ_DIR}/requests_upd ]; then
     touch ${PZ_DIR}/requests_upd
     echo "[*] installing new dependencies.."
     ${PZ_DIR}/venv/bin/pip install -U orjson requests &> /dev/null
@@ -36,21 +50,29 @@ fi
 
 if [ ! -x "$(command -v ldid)" ]; then
     echo "[*] installing ldid.."
-    sudo curl -so /usr/local/bin/ldid https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/ldid_${OS}_$ARCH
-    sudo chmod +x /usr/local/bin/ldid
+    sudo curl -so ${PATHPREFIX}/usr/local/bin/ldid https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/ldid_${OS}_$ARCH
+    sudo chmod +x ${PATHPREFIX}/usr/local/bin/ldid
 fi
 
 # install_name_tool and otool should only be installed here on linux
 if [ ! -x "$(command -v otool)" ]; then
     echo "[*] installing otool.."
-    sudo curl -so /usr/local/bin/otool https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/otool_$ARCH
-    sudo chmod +x /usr/local/bin/otool
+    sudo curl -so ${PATHPREFIX}/usr/local/bin/otool https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/otool_$ARCH
+    sudo chmod +x ${PATHPREFIX}/usr/local/bin/otool
 fi
 
 if [ ! -x "$(command -v install_name_tool)" ]; then
     echo "[*] installing install_name_tool.."
-    sudo curl -so /usr/local/bin/install_name_tool https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/install_name_tool_$ARCH
-    sudo chmod +x /usr/local/bin/install_name_tool
+    sudo curl -so ${PATHPREFIX}/usr/local/bin/install_name_tool https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/deps/install_name_tool_$ARCH
+    sudo chmod +x ${PATHPREFIX}/usr/local/bin/install_name_tool
+fi
+
+# lief is used on desktop, insert_dylib on iOS, so fetch that
+if [ "$OS" == "iPhone" ] && [ ! -x "$(command -v insert_dylib)" ]; then
+    # this *might* work? should test on both rootful/rootless
+    echo "[*] installing insert_dylib.."
+    sudo curl -so ${PATHPREFIX}/usr/local/bin/insert_dylib http://cf.yippee.love/insert_dylib
+    sudo chmod +x ${PATHPREFIX}/usr/local/bin/insert_dylib
 fi
 
 # create (or update) hidden dir
@@ -63,13 +85,13 @@ fi
 
 echo "[*] installing pyzule.."
 curl -so ~/.config/pyzule/version.json https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/version.json
-sudo rm /usr/local/bin/pyzule &> /dev/null  # yeah this is totally required leave me alone
-sudo curl -so /usr/local/bin/pyzule https://raw.githubusercontent.com/asdfzxcvbn/pyzule/main/pyzule.py
+sudo rm ${PATHPREFIX}/usr/local/bin/pyzule &> /dev/null  # yeah this is totally required leave me alone
+sudo curl -so ${PATHPREFIX}/usr/local/bin/pyzule ${PYZULEURL}
 if [ "$OS" == "Linux" ]; then
-    sudo sed -i "1s|.*|#\!${PZ_DIR}/venv/bin/python|" /usr/local/bin/pyzule
+    sudo sed -i "1s|.*|#\!${PZ_DIR}/venv/bin/python|" ${PATHPREFIX}/usr/local/bin/pyzule
 else
-    sudo sed -e "1s|.*|#\!${PZ_DIR}/venv/bin/python|" -i "" /usr/local/bin/pyzule  # bsd sed is broken asf
+    sudo sed -e "1s|.*|#\!${PZ_DIR}/venv/bin/python|" -i "" ${PATHPREFIX}/usr/local/bin/pyzule &> /dev/null  # bsd sed is broken asf
 fi
 echo "[*] fixed interpreter path!"
-sudo chmod +x /usr/local/bin/pyzule
+sudo chmod +x ${PATHPREFIX}/usr/local/bin/pyzule
 echo "[*] done!"
